Moved LoginForm component to src/components
