+++ src/pages/Login/index.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom'
-import LoginForm from '../../components/LoginForm/LoginForm.tsx'
+import LoginForm from '../../components/LoginForm'
 
 function LoginPage() {
   const navigate = useNavigate()
